refactor PostLayout and Handbook
