+++ src/mdxGlobalComponents.js
@@ -1,4 +1,4 @@
-// AUTO GENERATED FILE
+// AUTO GENERATED FILE 
 
 import { AboutAnchorScrollNavbar } from './components/About/AboutAnchorScrollNavbar'
 import { AboutBlog } from './components/About/AboutBlog'
@@ -140,142 +140,142 @@ import { VisitLibrary } from './components/VisitLibrary'
 import { ZoomImage } from './components/ZoomImage'
 
 export const shortcodes = {
-    AboutAnchorScrollNavbar,
-    AboutBlog,
-    AboutHero,
-    AboutInvestors,
-    AboutStory,
-    AboutTeam,
-    AboutTransparency,
-    Accordion,
-    AnchorScrollNavbar,
-    AnimateIntoView,
-    AnimatedBurger,
-    Apps,
-    AppsList,
-    ArrayCTA,
-    AshbyOpenRoles,
-    Banner,
-    BasicHedgehogImage,
-    BlockQuote,
-    Blog,
-    BlogFooter,
-    BorderWrapper,
-    Breadcrumbs,
-    CallToAction,
-    Caption,
-    Card,
-    Benefits,
-    CareersHero,
-    InterviewProcess,
-    OpenRoles,
-    Transparency,
-    WhyWereHere,
-    WorkingAtPostHog,
-    Checkbox,
-    Chip,
-    CodeBlock,
-    CommunityQuestions,
-    CompensationCalculator,
-    Contact,
-    Container,
-    ContributorCard,
-    ContributorSearch,
-    ContributorsChart,
-    CookieBanner,
-    Customers,
-    DarkModeToggle,
-    DemoLink,
-    DemoScheduler,
-    DeployOption,
-    DocsPageSurvey,
-    EU,
-    FeatureAvailability,
-    FeaturesSelfHost,
-    FloatedImage,
-    Footer,
-    FooterCTA,
-    FullWidthBorderSlider,
-    GDPRForm,
-    GitHubTooltip,
-    GithubIcon,
-    GlossaryElement,
-    Heading,
-    HiddenSection,
-    Home,
-    HostingOption,
-    Icons,
-    ImageBlock,
-    IngestionAppsList,
-    InlineCode,
-    LPCTA,
-    LandingPageCallToAction,
-    LibraryComparison,
-    LibraryFeatures,
-    Link,
-    Logo,
-    MainNav,
-    MdxAnchorHeaders,
-    Modal,
-    NewsletterForm,
-    NewsletterTutorial,
-    NextSteps,
-    NotFoundPage,
-    NotProductIcons,
-    OverflowXSection,
-    Popover,
-    PostCard,
-    PostLayout,
-    AllPlans,
-    AllPlansTest,
-    Calculator,
-    CloudVsSelfHost,
-    FAQs,
-    Features,
-    PlanComparison,
-    PlanComparisonTest,
-    PricingHero,
-    PricingSlider,
-    PricingTable,
-    ProductPillars,
-    Quote,
-    Savings,
-    SelfHost,
-    ProductAnchorNavbar,
-    ProductFeature,
-    ProductFeaturePlugin,
-    ProductFooter,
-    ProductIcons,
-    ProductPage,
-    ProductSectionHeader,
-    Questions,
-    Quote2,
-    RenderInClient,
-    Roadmap,
-    RocketHorizontalRule,
-    SectionLinks,
-    DesignedForYourStackBlock,
-    FeaturedSectionTextLeft,
-    FeaturedSectionTextRight,
-    FeaturedSectionTripleImage,
-    SignupCTA,
-    SiteSearchCallout,
-    SliderNav,
-    Spacer,
-    Spinner,
-    StarRepoButton,
-    StarUsBanner,
-    Structure,
-    Subscribe,
-    Tab,
-    TeamQuote,
-    Toast,
-    Toggle,
-    Tooltip,
-    TotalCountries,
-    TotalTeam,
-    Tutorials,
-    TutorialsSlider,
-    VisitLibrary,
-    ZoomImage,
+	AboutAnchorScrollNavbar,
+	AboutBlog,
+	AboutHero,
+	AboutInvestors,
+	AboutStory,
+	AboutTeam,
+	AboutTransparency,
+	Accordion,
+	AnchorScrollNavbar,
+	AnimateIntoView,
+	AnimatedBurger,
+	Apps,
+	AppsList,
+	ArrayCTA,
+	AshbyOpenRoles,
+	Banner,
+	BasicHedgehogImage,
+	BlockQuote,
+	Blog,
+	BlogFooter,
+	BorderWrapper,
+	Breadcrumbs,
+	CallToAction,
+	Caption,
+	Card,
+	Benefits,
+	CareersHero,
+	InterviewProcess,
+	OpenRoles,
+	Transparency,
+	WhyWereHere,
+	WorkingAtPostHog,
+	Checkbox,
+	Chip,
+	CodeBlock,
+	CommunityQuestions,
+	CompensationCalculator,
+	Contact,
+	Container,
+	ContributorCard,
+	ContributorSearch,
+	ContributorsChart,
+	CookieBanner,
+	Customers,
+	DarkModeToggle,
+	DemoLink,
+	DemoScheduler,
+	DeployOption,
+	DocsPageSurvey,
+	EU,
+	FeatureAvailability,
+	FeaturesSelfHost,
+	FloatedImage,
+	Footer,
+	FooterCTA,
+	FullWidthBorderSlider,
+	GDPRForm,
+	GitHubTooltip,
+	GithubIcon,
+	GlossaryElement,
+	Heading,
+	HiddenSection,
+	Home,
+	HostingOption,
+	Icons,
+	ImageBlock,
+	IngestionAppsList,
+	InlineCode,
+	LPCTA,
+	LandingPageCallToAction,
+	LibraryComparison,
+	LibraryFeatures,
+	Link,
+	Logo,
+	MainNav,
+	MdxAnchorHeaders,
+	Modal,
+	NewsletterForm,
+	NewsletterTutorial,
+	NextSteps,
+	NotFoundPage,
+	NotProductIcons,
+	OverflowXSection,
+	Popover,
+	PostCard,
+	PostLayout,
+	AllPlans,
+	AllPlansTest,
+	Calculator,
+	CloudVsSelfHost,
+	FAQs,
+	Features,
+	PlanComparison,
+	PlanComparisonTest,
+	PricingHero,
+	PricingSlider,
+	PricingTable,
+	ProductPillars,
+	Quote,
+	Savings,
+	SelfHost,
+	ProductAnchorNavbar,
+	ProductFeature,
+	ProductFeaturePlugin,
+	ProductFooter,
+	ProductIcons,
+	ProductPage,
+	ProductSectionHeader,
+	Questions,
+	Quote2,
+	RenderInClient,
+	Roadmap,
+	RocketHorizontalRule,
+	SectionLinks,
+	DesignedForYourStackBlock,
+	FeaturedSectionTextLeft,
+	FeaturedSectionTextRight,
+	FeaturedSectionTripleImage,
+	SignupCTA,
+	SiteSearchCallout,
+	SliderNav,
+	Spacer,
+	Spinner,
+	StarRepoButton,
+	StarUsBanner,
+	Structure,
+	Subscribe,
+	Tab,
+	TeamQuote,
+	Toast,
+	Toggle,
+	Tooltip,
+	TotalCountries,
+	TotalTeam,
+	Tutorials,
+	TutorialsSlider,
+	VisitLibrary,
+	ZoomImage
 }
